+++ clib/clamp.h
@@ -1,6 +1,6 @@
 #ifndef CLAMP_H
 #define CLAMP_H
 
-double clamp(float d, float min, float max);
+float clamp(float d, float min, float max);
 
 #endif
